fix job array scheduling anomaly
If some tasks of a job array are runnable and the meta-job array
record is not runable (e.g. held), the old logic could start a
runable task then try to start the non-runable meta-job, discover
it can not run, and set its reason to "BadConstraints".

Test case:
Make it so no jobs can start (partition stopped, slurmd down, etc.)
submit a job array
hold the job array
release the first two tasks of the job array
Make it so jobs can start