Add find_node_record2() function, no error if not found
Add a find_node_record2() function that differs from
find_node_record() in that it does not print an error if
a node name is not found. This is important for use with
the Cray capmc command, which generates information about
not only compute nodes, but also login and service nodes.
Using this function prevents a bunch of errors from the
power/cray plugin when trying to find these non-compute
nodes in the slurm configuration.