Insure prolog runs on job rqueue
Set the delay time for job requeue to the job credential lifetime (1200
second by default). This insures that prolog runs on every node when a
job is requeued. (This change will slow down launch of re-queued jobs).
Without this change, if a job is restated within 1200 seconds, the nodes
previously used would not run the prolog again, since the job ID is
still seen as active (from the previous execution). It is also advisable
to set the value of DEFAULT_EXPIRATION_WINDOW in src/common/slurm_cred.c
to the lowest value reasonable. We need to add a new configuration parameter
so this is easly changed in the future.