Reset job --mem=0 value
If a job is submitted with the option --mem=0, then once the job
is allocated resources, reset that value to the amount of memory
on the allocated node with the smallest amount of memory. This is
probably the best option for right now. The downside is that if
the job is requeued, its memory size will no longer be the total
memory on all nodes if the cluster is heterogeneous. On the upside,
some memory data is available for accounting purposes.
bug 1526