Limit user change of job priority
A non-administrator change to job priority will not be persistent except
for holding the job. User's wanting to change a job priority on a persistent
basis should reset it's "nice" value.
Without this change a user might lower a job's priority, but Slurm will
not lower it more based upon fair-share changes.