[PATCH 01/17] layouts_mgr: add shortkey field into layouts_keydef struct
The layouts_keydef struct only stored the associated key name prefixed by
the type of layouts (or mgr) inorder to have a unique identification of
all the keydefs in the central hash table.

To help accessing the original key name as specified in the layout conf
file, the shortkey field now stores a copy of the original key. This is
a short increase in the memory footprint of the layouts_keydef struct but
as this structure is only allocated once per key in a layout, the global
overhead is very small but provides an efficient way to retrieve the key
name.