[PATCH 02/17] layouts: initial layout packing/state_save logic addition
Add a function to help packing a particular layout into a Buffer as a set of
strings terminated by a "\0" string and representing the current expanded
form of the layout configuration as it could be described in a standard conf
file.

Use that function to provide a state save logic generating expanded conf files
in the state save location, one file per generated layout. Entities not linked
to a particular layout at save state time will not be saved as considered as
no longer necessary.

Note that the tree structure of the layouts are also saved in the packed
information as well as in the generated expanded file.

Note: fix a bug when freeing layouts in slurm_layouts_fini along the way.