[PATCH 03/17] layouts: initial layouts state recovery logic in slurmctld (entities data only)
Add a stage 1.1 in the loading of the layouts in order to recover the updated
entities prior to building the relationships among them (stage 2) using the
states saved in Slurm state save location for the layouts to load.

Information concerning the relations among entities provided by the 'Enclosed'
conf pragma are not taken into account for now when reading from layouts state
files. No layout plugins callbacks are called when doing that for now.