[PATCH 06/17] layouts: add layouts_update_layout() logic and factorize conf parsing calls
layouts_update_layout(...) now enables to process an incoming buffer containing
a contiguous set of packed strings and push that information into a particular
layout. Only the entities key/value pairs are updated for now.

layouts_p_update_done is a new callback that will be triggered with an array of
entities pointers when an update is performed on a particular layout. This
callback will help layout plugins to perform various consistency checks and/or
update any internal states that should be updated based on the modified entities
information.

No information concerning the udated key/value pairs are provided for now.
For finer-grained controls over what has changed, the plugins will have to use
both the parse_entity callback and the update_done callback to handle what has
changed and what should be rebuilt after the update for now.