Parse "#_*" as all tasks in job array
Modify slurmctld's parsing of a job_id string for the job_signal and
job_requeue calls to treat a job ID value of "#_*" as representing
all tasks in a job ID number "#". Previously treated as invalid input.

Also set the last_job_update time so that if a pending job is killed,
then that is reported immediately by "squeue -i#" (previously it
may keep reporting stale date.