Add thread-distribution option documentation
On nodes that support more than one thread per core, each thread can be configured as a separate logical CPU. The srun -m option allows users to specify the method for distributing CPUs across sockets (block, cyclic or fcyclic), but not across cores. The method for distributing CPUs across cores is hardcoded as block. The attached patches implement an extension to the -m option to allow users to request block, cyclic or fcyclic distribution of CPUs across cores. I have also implemented a wildcard character (*) to allow users to request the default value for each distribution type. The new distribution is supported by the task/cgroup plugin only, not task/affinity. The new syntax of the -m option is as follows:

-m, --distribution=
    *|block|cyclic|arbitrary|plane=<options> [:*|block|cyclic|fcyclic[:*|block|cyclic|fcyclic]][,Pack|NoPack]

The main reason for this change is to provide better support for the Xeon Phi architecture, which allows up to four threads per core.
This is dev-hyperthread1-1.patch - Documentation changes