Patch that adds new script and templates
I've attached a patch to add a new script and templates that create slurm.conf and gres.conf on a Cray SMW.

The script parses the output of xthwinv, combines nodes with the same hardware, and writes the output to slurm.conf. It also uses NodeName in gres.conf to allow having the same gres.conf file on all nodes.

I removed the settings in slurm.conf that we had commented out or set to the default.