Fix node state race condition, UNKNOWN->IDLE
If a job was running on the node when slurmctld restarted. The slurmd
would notify slurmctld when the job ended and slurmctld would change
its state from UNKNOWN to IDLE, at least if the job termination happened
prior to the slurmd being asked for configuration information. The
configuration information might then not be collected for some time.
I've modified the code to address this problem and try to collect
configuration information from every node after slurmctld startup,
eliminating this race condition.
bug 1805