[PATCH 1/4] Save free memory in the node infos
This commit adds a new field free_mem to the node_info struct
and augments the 15.08 RPC protocol to exchange this information
as part of the ping and registration status messages. The
free memory is taken from the sysinfo() information on systems
that support this function. Overall, the implementation sticks
to the way cpu_load (the 5 min load average) is handled.