Make PAM allocation and deallocation cleaner and safer
In pam_setup, correct comments for typos and to match the code.  On failure,
properly deallocate the PAM resources and ensure the PAM handle is null so as
to make calling pam_finish always safe.

In pam_finish, always try to deallocate all the PAM resources.  Also, ensure
the PAM handle is null so as to make calling pam_finish always safe.  This
allows for calling pam_finish multiple times without error.