Preserve job reason from start of scheduling cycle
At the start of a scheduling cycle, the job's "reason" field can
be cleared. If the scheduler fails to reach that job and set its
value to a new reason, the original reason was lost and the state
reports would report NoReason. This change saves the last reason
for a job being in a pending state and reports that value to the
user until we have a new valid reason for it still being in a
PENDING state.
bug 1919