Manual job resume bookkeeping fix
If a suspended job is manually resumed and gang scheduling is
configured, but no time slices are available for the job being
resumed, then just resume it without adding it to a time slice.
The jobs previously running on those nodes will be replaced with
new jobs as resources become available and the resumed job will
basically be treated like a stray job.
bug 2031