Streamline job array/file test logic
When slurmctld processes a SIGHUP or "scontrol reconfig" it
makes sure that every pending batch job as a directory containing
script and environment files, cancelling any job that lacks the
files and purging any files without a job. When there are a large
number of jobs (in any state) and a large number of files, the
nested loops to do this can run for a very long time (minutes).
This change streamlines the processing for job arrays, which now
avoid creating their own script and environment files by using that
of their job array master record.