update job/script file synchronization
When slurmctld is reconfigured, it makes sure that every pending
batch job has a script and if not mark the job as failed. If the
script is present with no job, then purge the file/directory.

This re-write of the logic takes advantage of hash tables to
improve the performance of the algorithm and is an improvement
on the logic in commit 862f3c8ee8e0500e76134cbce7a2722c621237a7
and will work with the version 15.08 code base as well
Time for for slurmctld reconfig with 110,000 jobs/files
37.8 sec with original code
2.2  sec with commit 862f3c8ee
0.8  sec with latest code
bug 2220