Rework messages when falling back to older directory format for job environment
Previously an error() would be logged when the attempt to open the job
script using the new directory format failed but the successive fallback to the
old directory structure was successful, leading to confusion when troubleshooting.

Move emitted warnings to debug(), and only error() after failing to open in both
directory structures. Add a note about backwards compatibility to both functions -
we cannot remove these fallbacks as directory structure for pending jobs does not
change on Slurm version update, and people may need to chain multiple version
update together to get to a current slurm version which would correctly update
slurmctld state files but leave pending jobs in the old directory structure.

Bug #2244.