Revert "Disable Cray NHC by default"
This reverts commit 6c21c4411dd88399b9ee8db35d34ce9d8748b202.

# Conflicts:
#	NEWS

This is being reverted because we have been informed turning off NHC by
default is not what Cray wants.  We hope for a day when this will be
the case and Slurm can run NHC on the compute node but until that day
this is the way we will handle it.