Fix for socket allocations and specialized cores
Previous logic would prevent allocation of sockets to a job unless the
entire socket was available. If there were any specialized cores, the
socket was treated as being not available and unusable. For example,
if a node had 2 sockets, then a job requesting 2 specialized cores
would reserve one core on each of the two sockets and render the job
not runnable.