Fix for srun signal handling threading problem
This is a revision to commit 1ed38f267e7044b249fb29f5d90ab9100d2a4edc
The root problem is that a pthread is passed an argument which is
a pointer to a variable on the stack. If that variable is over-written,
the signal number recieved will be garbage, and that bad signal
number will be interpretted by srun to possible abort the request.