Introduce siphash, and use it.
Siphash is a state of the art keyed hash function that is performance competitive with the usual non-cryptographic hash functions. It's used as the default hash function backing hash tables in e.g. Perl, Python, Rust, and so on. Here we initially use it for the gid cache hash table, and in the common xhash implementation.