Fix for job deadline with QOS MaxWall
If a job was submitted with a deadline and no time_limit or min_time,
but the system has a QOS MaxWall the job's time_limit would be set
to the QOS limit. Since there is no min_time specified, the QOS MaxWall
would be treated as a min and max time limit for the job and potentially
make the deadline impossible to satisfy. Now we set the min_time to
1 minute of there is a deadline, but no time_limit or min_time.