Burst_buffer paths fix
This add some additional logic to the commit made to version 15.08
as needed for operation with version 16.04. Specifically, once a
persistent burst buffer is created in versioin 16.04 the create
flag is cleared to prevent attempts at duplicate buffer create.
A new "use" persistent burst buffer is added for our needs (indicating
that a DataWarp "paths" operation is required). The first commit
is 905ac8509849a2dc331c09d529b63dcba45f7662