Fix task/cgroup linking
On some systems the char_to_val was not being put into the plugin,
resulting in the following error:
slurmstepd: [23.0]: symbol lookup error: /home/jette/SLURM/install_smd/lib/slurm/task_cgroup.so: undefined symbol: char_to_val
The problem was fixed by declaring the function "static". The
function was name was also updated with a leading "_" to indicate
the function is local to that module.