Fix build for bgq caused by redefinition of _GNU_SOURCE.
g++ always internally has a '#define _GNU_SOURCE 1'. The type difference
between that and '#define _GNU_SOURCE' which is used throughout Slurm
causes a compiler warning about redefinition, which breaks the build.

Adding this one #ifndef back solves this, as this is appears to be the
only header defining it that needs to build both with g++ and gcc.