Fix step cpus_per_task calculation
Correct logic that calculates a step's cpus_per_task allocation
on a heterogenous job allocation. Mixing a KNL with a Xeon resulted
in a count that was between the CPU count on the two node types
and invalid on the node with smaller CPU count (e.g. 272 CPUs on
KNL, 8 on Xeon, and 2 tasks, cpus_per_task = 140).