If no default account is given for a user when creating (only a list of accounts) no default account is printed, previously NULL was printed.
This is just not printing it, but whole function should probably be
revisited as the rigmarole can probably be avoided as we always know what
the default is going to be if none is specified (first off the list).

The problem with that though is if the user has been added to a cluster
already and they have a default, but then added to a new cluster where
they don't have a default.  In this case you want to keep the first
clusters default, but set the default for the second cluster.

Bug 2725