move buffer from stack to heap
The buffer to be used for reading the system /proc/*/stat files
is moved from the stack to the heap (i.e. malloc'ed memory) and
initialized to zero, and increased in size from 1k to 4k.
I don't see how this could make any difference, but this anomaly
was reported by valgrind.
bug 2234