Remove eligible_time field from jobcomp/elasticsearch.
Initially I wanted to record somehow the time a job has been in PD state
until it actually starts. To do so, I derived an eligible_time field which
was calculated as start_time - begin_time. This was not correct and a user
of the plugin reported that if the job remained a lot of time in PD state
the field underflowed. A more accurate approach would be calculating the
field as start_time - eligible_time (the actual Slurm eligible_time).
Since this field is not accessible within the log_record function which
receives a job_record struct as parameter, I've decided to just remove this
field from being logged.