Fix memory leaks for slurm_get_auth_info()
The slurm_get_auth_info() function returns an xmalloc'ed string
which few places were xfree'ing. The function in the past did
not xmalloc a value, but returned a pointer to a cached string.
When that was changed, the calls to the function were not all
changed, resulting in leaks if AuthInfo was defined in slurm.conf.