MYSQL - Sightly better logic if a job completion comes in with an end time of 0.
This might be the cause of run away jobs.  I couldn't see how an end_time
could be 0, but if it was it would just exit and never set time_end to
anything.  At least if it happens now we can have an idea that it is
possible and we will have an idea this is the place it happens.