Continuation of last commit.
We have decided to go back to the way 15.08 called NHC instead of calling
it first before sending a SIGKILL to the steps tasks.  With this patch we
only start the NHC early when we have to resend the SIGKILL for unkillable
processes.  This will hopefully get us the backtrace of the unkillable
processes which was the reason we did it this way in the first place :).