Various cleanup needed for extern step.  Continuation of commit 2fc0c860e64
What this does is set the state earlier to match a normal set.

Remove the unneeded _send_pending_exit_msgs.  There is only one task and
we have the message for it, so don't worry about that one.

Most important, wait for the other slurmstepd's to send their message,
otherwise they could be lost on the other end.