More on the others dealing with the extern cleanup.
What this does is treats the extern step like a normal step on exit.  It
doesn't appear the original code is needed anymore and this simplifies
the code.

The select_cray change is relevant since the add is needed only when
killing the step as that is the only place _internal_step_complete isn't
used.