Add SLURM_PENDING_STEP id so it won't be confused with SLURM_EXTERN_CONT.
In commit b8190e5d60dfca many places that were mean to be pending step ids
were changed to be extern_step id.  The main problem was when we came up
with the idea of the extern step we reused -1 (INFINITE) for the id.  So
pending steps also appeared to be extern steps as well.  Hopefully this
fixes the situation.

Bug 2907