Add boot request time to node record
This is required in order to make sure that a compute node reboots as
requested; even if the slurmctld daemon restarts. Previously if the
slurmctld daemon restarted, we couldn't tell if the compute node
rebooted before or after the request.
This commit expands upon the work of commit 4517c454be9d278b1021c93178914ecaebd97d1c
by adding the reboot request time to the node record, which is
saved and restored by slurmctld.
bug 3042